+++ pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
   flutter:
     sdk: flutter
   path_provider: ^1.0.0
+  universal_html: ^1.1.6
 
 dev_dependencies:
   flutter_test:
